refactor: rename footer component
